Scope integration tests to issue #35 only
Replace bundled issues_test.go with issue35_test.go to keep this branch
focused on CreateBulk.
